pyproject.toml
# pyproject.toml
[tool.pytest.ini_options]
testpaths = [
    "tests",
    # "pysui_flask/test/route"
]
pythonpath = ["pysui_flask"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pysui-flask"

description = "pysui-flask pysui flask REST api"
requires-python = ">=3.10"
keywords = [
    "software development kit",
    "sui",
    "blockchain",
    "utilities",
    "pysui",
    "flask",
    "framework",
]
license = { text = "Apache-2.0" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Programming Language :: Python",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    "pysui>=0.33.0",
    "Flask>=2.3.2",
    "flasgger>=0.9.7.2.dev2",
    "flask-marshmallow>=0.15.0",
    "apispec>=6.3.0",
    "python-dotenv>=1.0.0",
    "dataclasses-json",

]
dynamic = ["version", "readme"]


[project.scripts]

[tool.setuptools.packages.find]
exclude = [
    "out*",
    "temp*",
    "test*",
    "tools*",
    "build*",
    "doc*",
    "env*",
    "images*",
]


[tool.setuptools.dynamic]
version = { attr = "pysui_flask.version.__version__" }
readme = { file = ["README.rst"], content-type = "text/x-rst" }

[project.urls]
# changelog = ""
repository = "https://github.com/FrankC01/pysui-flask"
# issues = ""
